refactor: Evaluator를 생성하는 로직을 Factory로 분리
