+++ frontend/src/modules/evaluator/EvaluatorFactory.ts
@@ -1,0 +1,13 @@
+import EvalWorker from './eval.worker?worker';
+
+export default class EvaluatorFactory {
+  static createEvaluator() {
+    const evaluator = {
+      isIdle: true,
+      worker: new EvalWorker(),
+      currentTask: null,
+    };
+
+    return evaluator;
+  }
+}
